+++ sdk/communication/azure-communication-phonenumbers/azure/communication/phonenumber/_phone_number_administration_client.py
@@ -4,23 +4,16 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-from typing import Dict
-
 from azure.communication.phonenumber._generated.models import ReleaseStatus, CreateSearchOptions
-from azure.core.async_paging import AsyncItemPaged
 from azure.core.tracing.decorator import distributed_trace
-from azure.core.tracing.decorator_async import distributed_trace_async
-from azure.core.polling import AsyncLROPoller
+from azure.core.paging import ItemPaged
+from azure.core.polling import LROPoller
+from ._polling import ReleasePhoneNumberPolling, ReservePhoneNumberPolling, PurchaseReservationPolling
 
-from .._version import SDK_MONIKER
-from ._polling_async import ReleasePhoneNumberPollingAsync, \
-    ReservePhoneNumberPollingAsync, \
-    PurchaseReservationPollingAsync
-
-from .._generated.aio._phone_number_administration_service \
+from ._generated._phone_number_administration_service\
     import PhoneNumberAdministrationService as PhoneNumberAdministrationClientGen
 
-from .._generated.models import (
+from ._generated.models import (
     AcquiredPhoneNumbers,
     AreaCodes,
     LocationOptionsResponse,
@@ -38,9 +31,10 @@ from .._generated.models import (
     UpdatePhoneNumberCapabilitiesResponse
 )
 
-from .._shared.utils import parse_connection_str, get_authentication_policy
+from ._shared.utils import parse_connection_str, get_authentication_policy
+from ._version import SDK_MONIKER
 
-class PhoneNumberAdministrationClient(object):
+class PhoneNumbersAdministrationClient(object):
     """Azure Communication Services Phone Number Management client.
 
     :param str endpoint:
@@ -51,9 +45,9 @@ class PhoneNumberAdministrationClient(object):
     """
     def __init__(
             self,
-            endpoint,  # type: str
-            credential,  # type: str
-            **kwargs  # type: Any
+            endpoint, # type: str
+            credential, # type: str
+            **kwargs # type: Any
     ):
         # type: (...) -> None
         try:
@@ -69,7 +63,7 @@ class PhoneNumberAdministrationClient(object):
         self._endpoint = endpoint
         self._phone_number_administration_client = PhoneNumberAdministrationClientGen(
             self._endpoint,
-            authentication_policy=get_authentication_policy(endpoint, credential, is_async=True),
+            authentication_policy=get_authentication_policy(endpoint, credential),
             sdk_moniker=SDK_MONIKER,
             **kwargs)
 
@@ -77,13 +71,13 @@ class PhoneNumberAdministrationClient(object):
     def from_connection_string(
             cls, conn_str,  # type: str
             **kwargs  # type: Any
-    ):  # type: (...) -> PhoneNumberAdministrationClient
-        """Create PhoneNumberAdministrationClient from a Connection String.
-
+    ):
+        # type: (...) -> PhoneNumbersAdministrationClient
+        """Create PhoneNumbersAdministrationClient from a Connection String.
         :param str conn_str:
             A connection string to an Azure Communication Service resource.
-        :returns: Instance of PhoneNumberAdministrationClient.
-        :rtype: ~azure.communication.PhoneNumberAdministrationClient
+        :returns: Instance of PhoneNumbersAdministrationClient.
+        :rtype: ~azure.communication.PhoneNumbersAdministrationClient
         """
         endpoint, access_key = parse_connection_str(conn_str)
 
@@ -94,7 +88,7 @@ class PhoneNumberAdministrationClient(object):
             self,
             **kwargs  # type: Any
     ):
-        # type: (...) -> AsyncItemPaged[AcquiredPhoneNumbers]
+        # type: (...) -> ItemPaged[AcquiredPhoneNumbers]
         """Gets the list of the acquired phone numbers.
 
         :keyword str locale: A language-locale pairing which will be used to localise the names of countries.
@@ -103,14 +97,14 @@ class PhoneNumberAdministrationClient(object):
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.AcquiredPhoneNumbers]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.AcquiredPhoneNumbers]
         """
         return self._phone_number_administration_client.phone_number_administration.get_all_phone_numbers(
             **kwargs
         )
 
-    @distributed_trace_async
-    async def get_all_area_codes(
+    @distributed_trace
+    def get_all_area_codes(
             self,
             location_type,  # type: str
             country_code,  # type: str
@@ -130,15 +124,15 @@ class PhoneNumberAdministrationClient(object):
         Represents the underlying list of countries.
         :rtype: ~azure.communication.administration.AreaCodes
         """
-        return await self._phone_number_administration_client.phone_number_administration.get_all_area_codes(
-            location_type,
-            country_code,
-            phone_plan_id,
+        return self._phone_number_administration_client.phone_number_administration.get_all_area_codes(
+            location_type=location_type,
+            country_code=country_code,
+            phone_plan_id=phone_plan_id,
             **kwargs
         )
 
-    @distributed_trace_async
-    async def get_capabilities_update(
+    @distributed_trace
+    def get_capabilities_update(
         self,
         capabilities_update_id,  # type: str
         **kwargs  # type: Any
@@ -150,13 +144,13 @@ class PhoneNumberAdministrationClient(object):
         :type capabilities_update_id: str
         :rtype: ~azure.communication.administration.UpdatePhoneNumberCapabilitiesResponse
         """
-        return await self._phone_number_administration_client.phone_number_administration.get_capabilities_update(
+        return self._phone_number_administration_client.phone_number_administration.get_capabilities_update(
             capabilities_update_id,
             **kwargs
         )
 
-    @distributed_trace_async
-    async def update_capabilities(
+    @distributed_trace
+    def update_capabilities(
         self,
         phone_number_capabilities_update,  # type: Dict[str, NumberUpdateCapabilities]
         **kwargs  # type: Any
@@ -170,7 +164,7 @@ class PhoneNumberAdministrationClient(object):
          dict[str, ~azure.communication.administration.NumberUpdateCapabilities]
         :rtype: ~azure.communication.administration.UpdateNumberCapabilitiesResponse
         """
-        return await self._phone_number_administration_client.phone_number_administration.update_capabilities(
+        return self._phone_number_administration_client.phone_number_administration.update_capabilities(
             phone_number_capabilities_update,
             **kwargs
         )
@@ -180,10 +174,8 @@ class PhoneNumberAdministrationClient(object):
         self,
         **kwargs  # type: Any
     ):
-        # type: (...) -> AsyncItemPaged[PhoneNumberCountries]
+        # type: (...) -> ItemPaged[PhoneNumberCountries]
         """Gets a list of supported countries.
-
-        Gets a list of supported countries.
 
         :keyword str locale: A language-locale pairing which will be used to localise the names of countries.
         The default is "en-US".
@@ -191,18 +183,17 @@ class PhoneNumberAdministrationClient(object):
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :return: An iterator like instance of either PhoneNumberCountries or the result of cls(response)
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.PhoneNumberCountries]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.PhoneNumberCountries]
         """
         return self._phone_number_administration_client.phone_number_administration.get_all_supported_countries(
             **kwargs
         )
 
-    @distributed_trace_async
-    async def get_number_configuration(
-            self,
-            phone_number,  # type: str
-            **kwargs  # type: Any
+    @distributed_trace
+    def get_number_configuration(
+        self,
+        phone_number,  # type: str
+        **kwargs  # type: Any
     ):
         # type: (...) -> NumberConfigurationResponse
         """Endpoint for getting number configurations.
@@ -211,17 +202,17 @@ class PhoneNumberAdministrationClient(object):
         :type phone_number: str
         :rtype: ~azure.communication.administration.NumberConfigurationResponse
         """
-        return await self._phone_number_administration_client.phone_number_administration.get_number_configuration(
+        return self._phone_number_administration_client.phone_number_administration.get_number_configuration(
             phone_number,
             **kwargs
         )
 
-    @distributed_trace_async
-    async def configure_number(
-            self,
-            pstn_configuration,  # type: PstnConfiguration
-            phone_number,  # type: str
-            **kwargs  # type: Any
+    @distributed_trace
+    def configure_number(
+        self,
+        pstn_configuration,  # type: PstnConfiguration
+        phone_number,  # type: str
+        **kwargs  # type: Any
     ):
         # type: (...) -> None
         """Endpoint for configuring a pstn number.
@@ -232,17 +223,17 @@ class PhoneNumberAdministrationClient(object):
         :type phone_number: str
         :rtype: None
         """
-        return await self._phone_number_administration_client.phone_number_administration.configure_number(
+        return self._phone_number_administration_client.phone_number_administration.configure_number(
             pstn_configuration,
             phone_number,
             **kwargs
         )
 
-    @distributed_trace_async
-    async def unconfigure_number(
-            self,
-            phone_number,  # type: str
-            **kwargs  # type: Any
+    @distributed_trace
+    def unconfigure_number(
+        self,
+        phone_number,  # type: str
+        **kwargs  # type: Any
     ):
         # type: (...) -> None
         """Endpoint for unconfiguring a pstn number by removing the configuration.
@@ -251,18 +242,18 @@ class PhoneNumberAdministrationClient(object):
         :type phone_number: str
         :rtype: None
         """
-        return await self._phone_number_administration_client.phone_number_administration.unconfigure_number(
+        return self._phone_number_administration_client.phone_number_administration.unconfigure_number(
             phone_number,
             **kwargs
         )
 
     @distributed_trace
     def list_phone_plan_groups(
-            self,
-            country_code,  # type: str
-            **kwargs  # type: Any
+        self,
+        country_code,  # type: str
+        **kwargs  # type: Any
     ):
-        # type: (...) -> AsyncItemPaged[PhonePlanGroups]
+        # type: (...) -> ItemPaged[PhonePlanGroups]
         """Gets a list of phone plan groups for the given country.
 
         :param country_code: The ISO 3166-2 country code.
@@ -275,7 +266,7 @@ class PhoneNumberAdministrationClient(object):
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.PhonePlanGroups]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.PhonePlanGroups]
         """
         return self._phone_number_administration_client.phone_number_administration.get_phone_plan_groups(
             country_code,
@@ -284,12 +275,12 @@ class PhoneNumberAdministrationClient(object):
 
     @distributed_trace
     def list_phone_plans(
-            self,
-            country_code,  # type: str
-            phone_plan_group_id,  # type: str
-            **kwargs  # type: Any
+        self,
+        country_code,  # type: str
+        phone_plan_group_id,  # type: str
+        **kwargs  # type: Any
     ):
-        # type: (...) -> AsyncItemPaged[PhonePlansResponse]
+        # type: (...) -> ItemPaged[PhonePlansResponse]
         """Gets a list of phone plans for a phone plan group.
 
         :param country_code: The ISO 3166-2 country code.
@@ -302,7 +293,7 @@ class PhoneNumberAdministrationClient(object):
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.PhonePlansResponse]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.PhonePlansResponse]
         """
         return self._phone_number_administration_client.phone_number_administration.get_phone_plans(
             country_code,
@@ -310,13 +301,13 @@ class PhoneNumberAdministrationClient(object):
             **kwargs
         )
 
-    @distributed_trace_async
-    async def get_phone_plan_location_options(
-            self,
-            country_code,  # type: str
-            phone_plan_group_id,  # type: str
-            phone_plan_id,  # type: str
-            **kwargs  # type: Any
+    @distributed_trace
+    def get_phone_plan_location_options(
+        self,
+        country_code,  # type: str
+        phone_plan_group_id,  # type: str
+        phone_plan_id,  # type: str
+        **kwargs  # type: Any
     ):
         # type: (...) -> LocationOptionsResponse
         """Gets a list of location options for a phone plan.
@@ -333,19 +324,17 @@ class PhoneNumberAdministrationClient(object):
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype:
-        ~azure.communication.administration.LocationOptionsResponse
+        :rtype: ~azure.communication.administration.LocationOptionsResponse
         """
-        return await self._phone_number_administration_client.\
-            phone_number_administration.get_phone_plan_location_options(
-            country_code,
-            phone_plan_group_id,
-            phone_plan_id,
+        return self._phone_number_administration_client.phone_number_administration.get_phone_plan_location_options(
+            country_code=country_code,
+            phone_plan_group_id=phone_plan_group_id,
+            phone_plan_id=phone_plan_id,
             **kwargs
         )
 
-    @distributed_trace_async
-    async def get_release_by_id(
+    @distributed_trace
+    def get_release_by_id(
             self,
             release_id,  # type: str
             **kwargs  # type: Any
@@ -357,17 +346,17 @@ class PhoneNumberAdministrationClient(object):
         :type release_id: str
         :rtype: ~azure.communication.administration.PhoneNumberRelease
         """
-        return await self._phone_number_administration_client.phone_number_administration.get_release_by_id(
+        return self._phone_number_administration_client.phone_number_administration.get_release_by_id(
             release_id,
             **kwargs
         )
 
-    @distributed_trace_async
-    async def begin_release_phone_numbers(
+    @distributed_trace
+    def begin_release_phone_numbers(
             self,
             **kwargs  # type: Any
     ):
-        # type: (...) -> AsyncLROPoller[PhoneNumberRelease]
+        # type: (...) -> LROPoller[PhoneNumberRelease]
         """Begins creating a release for the given phone numbers.
         Caller must provide either phone_numbers, or continuation_token keywords to use the method.
         If both phone_numbers and continuation_token are specified, only continuation_token will be used to
@@ -375,11 +364,11 @@ class PhoneNumberAdministrationClient(object):
 
         :keyword list[str] phone_numbers: The list of phone numbers in the release request.
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.communication.administration.PhoneNumberRelease]
+        :rtype: ~azure.core.polling.LROPoller[~azure.communication.administration.PhoneNumberRelease]
         """
         cont_token = kwargs.pop('continuation_token', None)  # type: Optional[str]
 
-        release_polling = ReleasePhoneNumberPollingAsync(
+        release_polling = ReleasePhoneNumberPolling(
             is_terminated=lambda status: status in [
                 ReleaseStatus.Complete,
                 ReleaseStatus.Failed,
@@ -388,7 +377,7 @@ class PhoneNumberAdministrationClient(object):
         )
 
         if cont_token is not None:
-            return AsyncLROPoller.from_continuation_token(
+            return LROPoller.from_continuation_token(
                 polling_method=release_polling,
                 continuation_token=cont_token,
                 client=self._phone_number_administration_client.phone_number_administration
@@ -397,14 +386,16 @@ class PhoneNumberAdministrationClient(object):
         if "phone_numbers" not in kwargs:
             raise ValueError("Either kwarg 'phone_numbers' or 'continuation_token' needs to be specified")
 
-        create_release_response = await self._phone_number_administration_client.\
+        create_release_response = self._phone_number_administration_client.\
             phone_number_administration.release_phone_numbers(
                 **kwargs
-            )
-        initial_state = await self._phone_number_administration_client.phone_number_administration.get_release_by_id(
+        )
+
+        initial_state = self._phone_number_administration_client.phone_number_administration.get_release_by_id(
             release_id=create_release_response.release_id
         )
-        return AsyncLROPoller(client=self._phone_number_administration_client.phone_number_administration,
+
+        return LROPoller(client=self._phone_number_administration_client.phone_number_administration,
                          initial_response=initial_state,
                          deserialization_callback=None,
                          polling_method=release_polling)
@@ -414,24 +405,24 @@ class PhoneNumberAdministrationClient(object):
             self,
             **kwargs  # type: Any
     ):
-        # type: (...) -> AsyncItemPaged[PhoneNumberEntities]
+        # type: (...) -> ItemPaged[PhoneNumberEntities]
         """Gets a list of all releases.
 
         :keyword int skip: An optional parameter for how many entries to skip, for pagination purposes.
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.PhoneNumberEntities]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.PhoneNumberEntities]
         """
         return self._phone_number_administration_client.phone_number_administration.get_all_releases(
             **kwargs
         )
 
-    @distributed_trace_async
-    async def get_reservation_by_id(
-            self,
-            reservation_id,  # type: str
-            **kwargs  # type: Any
+    @distributed_trace
+    def get_reservation_by_id(
+        self,
+        reservation_id,  # type: str
+        **kwargs  # type: Any
     ):
         # type: (...) -> PhoneNumberReservation
         """Get reservation by reservation id.
@@ -440,18 +431,17 @@ class PhoneNumberAdministrationClient(object):
         :type reservation_id: str
         :rtype: ~azure.communication.administration.PhoneNumberReservation
         """
-        return await self._phone_number_administration_client.phone_number_administration.get_search_by_id(
+        return self._phone_number_administration_client.phone_number_administration.get_search_by_id(
             search_id=reservation_id,
             **kwargs
         )
 
-
-    @distributed_trace_async
-    async def begin_reserve_phone_numbers(
-            self,
-            **kwargs  # type: Any
+    @distributed_trace
+    def begin_reserve_phone_numbers(
+        self,
+        **kwargs  # type: Any
     ):
-        # type: (...) -> AsyncLROPoller[PhoneNumberReservation]
+        # type: (...) -> LROPoller[PhoneNumberReservation]
         """Begins creating a phone number search to reserve phone numbers.
         Caller must provide one of the following:
          (1) all of keywords display_name, description, phone_plan_ids, area_code, quantity if all the phone plans
@@ -469,11 +459,11 @@ class PhoneNumberAdministrationClient(object):
         :keyword list[~azure.communication.administration.models.LocationOptionsDetails] location_options:
             the location options of the search.
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.communication.administration.PhoneNumberReservation]
+        :rtype: ~azure.core.polling.LROPoller[~azure.communication.administration.PhoneNumberReservation]
         """
         cont_token = kwargs.pop('continuation_token', None)  # type: Optional[str]
 
-        reservation_polling = ReservePhoneNumberPollingAsync(
+        reservation_polling = ReservePhoneNumberPolling(
             is_terminated=lambda status: status in [
                 SearchStatus.Reserved,
                 SearchStatus.Expired,
@@ -484,7 +474,7 @@ class PhoneNumberAdministrationClient(object):
         )
 
         if cont_token is not None:
-            return AsyncLROPoller.from_continuation_token(
+            return LROPoller.from_continuation_token(
                 polling_method=reservation_polling,
                 continuation_token=cont_token,
                 client=self._phone_number_administration_client.phone_number_administration
@@ -508,15 +498,16 @@ class PhoneNumberAdministrationClient(object):
         if 'location_options' in kwargs:
             reservation_options.location_options = kwargs.pop('location_options')
 
-        create_reservation_response = await self._phone_number_administration_client.\
+        create_reservation_response = self._phone_number_administration_client.\
             phone_number_administration.create_search(
                 body=reservation_options,
                 **kwargs
         )
-        initial_state = await self._phone_number_administration_client.phone_number_administration.get_search_by_id(
+
+        initial_state = self._phone_number_administration_client.phone_number_administration.get_search_by_id(
             search_id=create_reservation_response.search_id
         )
-        return AsyncLROPoller(client=self._phone_number_administration_client.phone_number_administration,
+        return LROPoller(client=self._phone_number_administration_client.phone_number_administration,
                          initial_response=initial_state,
                          deserialization_callback=None,
                          polling_method=reservation_polling)
@@ -526,24 +517,24 @@ class PhoneNumberAdministrationClient(object):
             self,
             **kwargs  # type: Any
     ):
-        # type: (...) -> AsyncItemPaged[PhoneNumberEntities]
+        # type: (...) -> ItemPaged[PhoneNumberEntities]
         """Gets a list of all reservations.
 
         :keyword int skip: An optional parameter for how many entries to skip, for pagination purposes.
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.PhoneNumberEntities]
+        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.PhoneNumberEntities]
         """
         return self._phone_number_administration_client.phone_number_administration.get_all_searches(
             **kwargs
         )
 
-    @distributed_trace_async
-    async def cancel_reservation(
-            self,
-            reservation_id,  # type: str
-            **kwargs  # type: Any
+    @distributed_trace
+    def cancel_reservation(
+        self,
+        reservation_id,  # type: str
+        **kwargs  # type: Any
     ):
         # type: (...) -> None
         """Cancels the reservation. This means existing numbers in the reservation will be made available.
@@ -552,29 +543,28 @@ class PhoneNumberAdministrationClient(object):
         :type reservation_id: str
         :rtype: None
         """
-        return await self._phone_number_administration_client.phone_number_administration.cancel_search(
+        return self._phone_number_administration_client.phone_number_administration.cancel_search(
             search_id=reservation_id,
             **kwargs
         )
 
-    @distributed_trace_async
-    async def begin_purchase_reservation(
-            self,
-            **kwargs  # type: Any
+    @distributed_trace
+    def begin_purchase_reservation(
+        self,
+        **kwargs  # type: Any
     ):
-
-        # type: (...) -> AsyncLROPoller[PhoneNumberReservation]
+        # type: (...) -> LROPoller[PhoneNumberReservation]
         """Begins purchase the reserved phone numbers of a phone number search.
         Caller must provide either reservation_id, or continuation_token keywords to use the method.
         If both reservation_id and continuation_token are specified, only continuation_token will be used to
         restart a poller from a saved state, and keyword reservation_id will be ignored.
         :keyword str reservation_id: The reservation id to be purchased.
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.communication.administration.PhoneNumberReservation]
+        :rtype: ~azure.core.polling.LROPoller[~azure.communication.administration.PhoneNumberReservation]
         """
         cont_token = kwargs.pop('continuation_token', None)  # type: Optional[str]
 
-        reservation_polling = PurchaseReservationPollingAsync(
+        reservation_polling = PurchaseReservationPolling(
             is_terminated=lambda status: status in [
                 SearchStatus.Success,
                 SearchStatus.Expired,
@@ -584,37 +574,25 @@ class PhoneNumberAdministrationClient(object):
         )
 
         if cont_token is not None:
-            return AsyncLROPoller.from_continuation_token(
+            return LROPoller.from_continuation_token(
                 polling_method=reservation_polling,
                 continuation_token=cont_token,
                 client=self._phone_number_administration_client.phone_number_administration
             )
 
-        reservation_id = kwargs.pop('reservation_id', None)  # type: str
-        if reservation_id is None:
+        if "reservation_id" not in kwargs:
             raise ValueError("Either kwarg 'reservation_id' or 'continuation_token' needs to be specified")
 
-        await self._phone_number_administration_client.phone_number_administration.purchase_search(
-            reservation_id,
+        reservation_id = kwargs.pop('reservation_id')  # type: str
+
+        self._phone_number_administration_client.phone_number_administration.purchase_search(
+            search_id=reservation_id,
             **kwargs
         )
-        initial_state = await self._phone_number_administration_client.phone_number_administration.get_search_by_id(
+        initial_state = self._phone_number_administration_client.phone_number_administration.get_search_by_id(
             search_id=reservation_id
         )
-        return AsyncLROPoller(client=self._phone_number_administration_client.phone_number_administration,
+        return LROPoller(client=self._phone_number_administration_client.phone_number_administration,
                          initial_response=initial_state,
                          deserialization_callback=None,
                          polling_method=reservation_polling)
-
-    async def __aenter__(self) -> "PhoneNumberAdministrationClient":
-        await self._phone_number_administration_client.__aenter__()
-        return self
-
-    async def __aexit__(self, *args: "Any") -> None:
-        await self.close()
-
-    async def close(self) -> None:
-        """Close the :class:
-        `~azure.communication.administration.aio.PhoneNumberAdministrationClient` session.
-        """
-        await self._phone_number_administration_client.__aexit__()