+++ sdk/communication/azure-communication-phonenumbers/azure/communication/phonenumber/_polling.py
@@ -69,7 +69,7 @@ class PhoneNumberBasePolling(PollingMethod):
 
     @classmethod
     def from_continuation_token(cls, continuation_token, client, **kwargs):  # pylint: disable=W0221
-        # type(str, PhoneNumberAdministrationClient, Any) -> Tuple
+        # type(str, PhoneNumbersAdministrationClient, Any) -> Tuple
         initial_response = pickle.loads(base64.b64decode(continuation_token))  # nosec
         return client, initial_response, None