+++ sdk/communication/azure-communication-phonenumbers/azure/communication/phonenumber/aio/_phone_number_administration_client_async.py
@@ -4,16 +4,23 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-from azure.communication.phonenumber._generated.models import ReleaseStatus, CreateSearchOptions
-from azure.core.tracing.decorator import distributed_trace
-from azure.core.paging import ItemPaged
-from azure.core.polling import LROPoller
-from ._polling import ReleasePhoneNumberPolling, ReservePhoneNumberPolling, PurchaseReservationPolling
+from typing import Dict
 
-from ._generated._phone_number_administration_service\
+from azure.communication.phonenumber._generated.models import ReleaseStatus, CreateSearchOptions
+from azure.core.async_paging import AsyncItemPaged
+from azure.core.tracing.decorator import distributed_trace
+from azure.core.tracing.decorator_async import distributed_trace_async
+from azure.core.polling import AsyncLROPoller
+
+from .._version import SDK_MONIKER
+from ._polling_async import ReleasePhoneNumberPollingAsync, \
+    ReservePhoneNumberPollingAsync, \
+    PurchaseReservationPollingAsync
+
+from .._generated.aio._phone_number_administration_service \
     import PhoneNumberAdministrationService as PhoneNumberAdministrationClientGen
 
-from ._generated.models import (
+from .._generated.models import (
     AcquiredPhoneNumbers,
     AreaCodes,
     LocationOptionsResponse,
@@ -31,8 +38,7 @@ from ._generated.models import (
     UpdatePhoneNumberCapabilitiesResponse
 )
 
-from ._shared.utils import parse_connection_str, get_authentication_policy
-from ._version import SDK_MONIKER
+from .._shared.utils import parse_connection_str, get_authentication_policy
 
 class PhoneNumberAdministrationClient(object):
     """Azure Communication Services Phone Number Management client.
@@ -45,9 +51,9 @@ class PhoneNumberAdministrationClient(object):
     """
     def __init__(
             self,
-            endpoint, # type: str
-            credential, # type: str
-            **kwargs # type: Any
+            endpoint,  # type: str
+            credential,  # type: str
+            **kwargs  # type: Any
     ):
         # type: (...) -> None
         try:
@@ -63,7 +69,7 @@ class PhoneNumberAdministrationClient(object):
         self._endpoint = endpoint
         self._phone_number_administration_client = PhoneNumberAdministrationClientGen(
             self._endpoint,
-            authentication_policy=get_authentication_policy(endpoint, credential),
+            authentication_policy=get_authentication_policy(endpoint, credential, is_async=True),
             sdk_moniker=SDK_MONIKER,
             **kwargs)
 
@@ -71,13 +77,13 @@ class PhoneNumberAdministrationClient(object):
     def from_connection_string(
             cls, conn_str,  # type: str
             **kwargs  # type: Any
-    ):
-        # type: (...) -> PhoneNumberAdministrationClient
-        """Create PhoneNumberAdministrationClient from a Connection String.
+    ):  # type: (...) -> PhoneNumberAdministrationClient
+        """Create PhoneNumbersAdministrationClient from a Connection String.
+
         :param str conn_str:
             A connection string to an Azure Communication Service resource.
-        :returns: Instance of PhoneNumberAdministrationClient.
-        :rtype: ~azure.communication.PhoneNumberAdministrationClient
+        :returns: Instance of PhoneNumbersAdministrationClient.
+        :rtype: ~azure.communication.PhoneNumbersAdministrationClient
         """
         endpoint, access_key = parse_connection_str(conn_str)
 
@@ -88,7 +94,7 @@ class PhoneNumberAdministrationClient(object):
             self,
             **kwargs  # type: Any
     ):
-        # type: (...) -> ItemPaged[AcquiredPhoneNumbers]
+        # type: (...) -> AsyncItemPaged[AcquiredPhoneNumbers]
         """Gets the list of the acquired phone numbers.
 
         :keyword str locale: A language-locale pairing which will be used to localise the names of countries.
@@ -97,14 +103,14 @@ class PhoneNumberAdministrationClient(object):
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.AcquiredPhoneNumbers]
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.AcquiredPhoneNumbers]
         """
         return self._phone_number_administration_client.phone_number_administration.get_all_phone_numbers(
             **kwargs
         )
 
-    @distributed_trace
-    def get_all_area_codes(
+    @distributed_trace_async
+    async def get_all_area_codes(
             self,
             location_type,  # type: str
             country_code,  # type: str
@@ -124,15 +130,15 @@ class PhoneNumberAdministrationClient(object):
         Represents the underlying list of countries.
         :rtype: ~azure.communication.administration.AreaCodes
         """
-        return self._phone_number_administration_client.phone_number_administration.get_all_area_codes(
-            location_type=location_type,
-            country_code=country_code,
-            phone_plan_id=phone_plan_id,
+        return await self._phone_number_administration_client.phone_number_administration.get_all_area_codes(
+            location_type,
+            country_code,
+            phone_plan_id,
             **kwargs
         )
 
-    @distributed_trace
-    def get_capabilities_update(
+    @distributed_trace_async
+    async def get_capabilities_update(
         self,
         capabilities_update_id,  # type: str
         **kwargs  # type: Any
@@ -144,13 +150,13 @@ class PhoneNumberAdministrationClient(object):
         :type capabilities_update_id: str
         :rtype: ~azure.communication.administration.UpdatePhoneNumberCapabilitiesResponse
         """
-        return self._phone_number_administration_client.phone_number_administration.get_capabilities_update(
+        return await self._phone_number_administration_client.phone_number_administration.get_capabilities_update(
             capabilities_update_id,
             **kwargs
         )
 
-    @distributed_trace
-    def update_capabilities(
+    @distributed_trace_async
+    async def update_capabilities(
         self,
         phone_number_capabilities_update,  # type: Dict[str, NumberUpdateCapabilities]
         **kwargs  # type: Any
@@ -164,7 +170,7 @@ class PhoneNumberAdministrationClient(object):
          dict[str, ~azure.communication.administration.NumberUpdateCapabilities]
         :rtype: ~azure.communication.administration.UpdateNumberCapabilitiesResponse
         """
-        return self._phone_number_administration_client.phone_number_administration.update_capabilities(
+        return await self._phone_number_administration_client.phone_number_administration.update_capabilities(
             phone_number_capabilities_update,
             **kwargs
         )
@@ -174,8 +180,10 @@ class PhoneNumberAdministrationClient(object):
         self,
         **kwargs  # type: Any
     ):
-        # type: (...) -> ItemPaged[PhoneNumberCountries]
+        # type: (...) -> AsyncItemPaged[PhoneNumberCountries]
         """Gets a list of supported countries.
+
+        Gets a list of supported countries.
 
         :keyword str locale: A language-locale pairing which will be used to localise the names of countries.
         The default is "en-US".
@@ -183,17 +191,18 @@ class PhoneNumberAdministrationClient(object):
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.PhoneNumberCountries]
+        :return: An iterator like instance of either PhoneNumberCountries or the result of cls(response)
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.PhoneNumberCountries]
         """
         return self._phone_number_administration_client.phone_number_administration.get_all_supported_countries(
             **kwargs
         )
 
-    @distributed_trace
-    def get_number_configuration(
-        self,
-        phone_number,  # type: str
-        **kwargs  # type: Any
+    @distributed_trace_async
+    async def get_number_configuration(
+            self,
+            phone_number,  # type: str
+            **kwargs  # type: Any
     ):
         # type: (...) -> NumberConfigurationResponse
         """Endpoint for getting number configurations.
@@ -202,17 +211,17 @@ class PhoneNumberAdministrationClient(object):
         :type phone_number: str
         :rtype: ~azure.communication.administration.NumberConfigurationResponse
         """
-        return self._phone_number_administration_client.phone_number_administration.get_number_configuration(
+        return await self._phone_number_administration_client.phone_number_administration.get_number_configuration(
             phone_number,
             **kwargs
         )
 
-    @distributed_trace
-    def configure_number(
-        self,
-        pstn_configuration,  # type: PstnConfiguration
-        phone_number,  # type: str
-        **kwargs  # type: Any
+    @distributed_trace_async
+    async def configure_number(
+            self,
+            pstn_configuration,  # type: PstnConfiguration
+            phone_number,  # type: str
+            **kwargs  # type: Any
     ):
         # type: (...) -> None
         """Endpoint for configuring a pstn number.
@@ -223,17 +232,17 @@ class PhoneNumberAdministrationClient(object):
         :type phone_number: str
         :rtype: None
         """
-        return self._phone_number_administration_client.phone_number_administration.configure_number(
+        return await self._phone_number_administration_client.phone_number_administration.configure_number(
             pstn_configuration,
             phone_number,
             **kwargs
         )
 
-    @distributed_trace
-    def unconfigure_number(
-        self,
-        phone_number,  # type: str
-        **kwargs  # type: Any
+    @distributed_trace_async
+    async def unconfigure_number(
+            self,
+            phone_number,  # type: str
+            **kwargs  # type: Any
     ):
         # type: (...) -> None
         """Endpoint for unconfiguring a pstn number by removing the configuration.
@@ -242,18 +251,18 @@ class PhoneNumberAdministrationClient(object):
         :type phone_number: str
         :rtype: None
         """
-        return self._phone_number_administration_client.phone_number_administration.unconfigure_number(
+        return await self._phone_number_administration_client.phone_number_administration.unconfigure_number(
             phone_number,
             **kwargs
         )
 
     @distributed_trace
     def list_phone_plan_groups(
-        self,
-        country_code,  # type: str
-        **kwargs  # type: Any
+            self,
+            country_code,  # type: str
+            **kwargs  # type: Any
     ):
-        # type: (...) -> ItemPaged[PhonePlanGroups]
+        # type: (...) -> AsyncItemPaged[PhonePlanGroups]
         """Gets a list of phone plan groups for the given country.
 
         :param country_code: The ISO 3166-2 country code.
@@ -266,7 +275,7 @@ class PhoneNumberAdministrationClient(object):
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.PhonePlanGroups]
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.PhonePlanGroups]
         """
         return self._phone_number_administration_client.phone_number_administration.get_phone_plan_groups(
             country_code,
@@ -275,12 +284,12 @@ class PhoneNumberAdministrationClient(object):
 
     @distributed_trace
     def list_phone_plans(
-        self,
-        country_code,  # type: str
-        phone_plan_group_id,  # type: str
-        **kwargs  # type: Any
+            self,
+            country_code,  # type: str
+            phone_plan_group_id,  # type: str
+            **kwargs  # type: Any
     ):
-        # type: (...) -> ItemPaged[PhonePlansResponse]
+        # type: (...) -> AsyncItemPaged[PhonePlansResponse]
         """Gets a list of phone plans for a phone plan group.
 
         :param country_code: The ISO 3166-2 country code.
@@ -293,7 +302,7 @@ class PhoneNumberAdministrationClient(object):
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.PhonePlansResponse]
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.PhonePlansResponse]
         """
         return self._phone_number_administration_client.phone_number_administration.get_phone_plans(
             country_code,
@@ -301,13 +310,13 @@ class PhoneNumberAdministrationClient(object):
             **kwargs
         )
 
-    @distributed_trace
-    def get_phone_plan_location_options(
-        self,
-        country_code,  # type: str
-        phone_plan_group_id,  # type: str
-        phone_plan_id,  # type: str
-        **kwargs  # type: Any
+    @distributed_trace_async
+    async def get_phone_plan_location_options(
+            self,
+            country_code,  # type: str
+            phone_plan_group_id,  # type: str
+            phone_plan_id,  # type: str
+            **kwargs  # type: Any
     ):
         # type: (...) -> LocationOptionsResponse
         """Gets a list of location options for a phone plan.
@@ -324,17 +333,19 @@ class PhoneNumberAdministrationClient(object):
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.communication.administration.LocationOptionsResponse
+        :rtype:
+        ~azure.communication.administration.LocationOptionsResponse
         """
-        return self._phone_number_administration_client.phone_number_administration.get_phone_plan_location_options(
-            country_code=country_code,
-            phone_plan_group_id=phone_plan_group_id,
-            phone_plan_id=phone_plan_id,
+        return await self._phone_number_administration_client.\
+            phone_number_administration.get_phone_plan_location_options(
+            country_code,
+            phone_plan_group_id,
+            phone_plan_id,
             **kwargs
         )
 
-    @distributed_trace
-    def get_release_by_id(
+    @distributed_trace_async
+    async def get_release_by_id(
             self,
             release_id,  # type: str
             **kwargs  # type: Any
@@ -346,17 +357,17 @@ class PhoneNumberAdministrationClient(object):
         :type release_id: str
         :rtype: ~azure.communication.administration.PhoneNumberRelease
         """
-        return self._phone_number_administration_client.phone_number_administration.get_release_by_id(
+        return await self._phone_number_administration_client.phone_number_administration.get_release_by_id(
             release_id,
             **kwargs
         )
 
-    @distributed_trace
-    def begin_release_phone_numbers(
+    @distributed_trace_async
+    async def begin_release_phone_numbers(
             self,
             **kwargs  # type: Any
     ):
-        # type: (...) -> LROPoller[PhoneNumberRelease]
+        # type: (...) -> AsyncLROPoller[PhoneNumberRelease]
         """Begins creating a release for the given phone numbers.
         Caller must provide either phone_numbers, or continuation_token keywords to use the method.
         If both phone_numbers and continuation_token are specified, only continuation_token will be used to
@@ -364,11 +375,11 @@ class PhoneNumberAdministrationClient(object):
 
         :keyword list[str] phone_numbers: The list of phone numbers in the release request.
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :rtype: ~azure.core.polling.LROPoller[~azure.communication.administration.PhoneNumberRelease]
+        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.communication.administration.PhoneNumberRelease]
         """
         cont_token = kwargs.pop('continuation_token', None)  # type: Optional[str]
 
-        release_polling = ReleasePhoneNumberPolling(
+        release_polling = ReleasePhoneNumberPollingAsync(
             is_terminated=lambda status: status in [
                 ReleaseStatus.Complete,
                 ReleaseStatus.Failed,
@@ -377,7 +388,7 @@ class PhoneNumberAdministrationClient(object):
         )
 
         if cont_token is not None:
-            return LROPoller.from_continuation_token(
+            return AsyncLROPoller.from_continuation_token(
                 polling_method=release_polling,
                 continuation_token=cont_token,
                 client=self._phone_number_administration_client.phone_number_administration
@@ -386,16 +397,14 @@ class PhoneNumberAdministrationClient(object):
         if "phone_numbers" not in kwargs:
             raise ValueError("Either kwarg 'phone_numbers' or 'continuation_token' needs to be specified")
 
-        create_release_response = self._phone_number_administration_client.\
+        create_release_response = await self._phone_number_administration_client.\
             phone_number_administration.release_phone_numbers(
                 **kwargs
-        )
-
-        initial_state = self._phone_number_administration_client.phone_number_administration.get_release_by_id(
+            )
+        initial_state = await self._phone_number_administration_client.phone_number_administration.get_release_by_id(
             release_id=create_release_response.release_id
         )
-
-        return LROPoller(client=self._phone_number_administration_client.phone_number_administration,
+        return AsyncLROPoller(client=self._phone_number_administration_client.phone_number_administration,
                          initial_response=initial_state,
                          deserialization_callback=None,
                          polling_method=release_polling)
@@ -405,24 +414,24 @@ class PhoneNumberAdministrationClient(object):
             self,
             **kwargs  # type: Any
     ):
-        # type: (...) -> ItemPaged[PhoneNumberEntities]
+        # type: (...) -> AsyncItemPaged[PhoneNumberEntities]
         """Gets a list of all releases.
 
         :keyword int skip: An optional parameter for how many entries to skip, for pagination purposes.
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.PhoneNumberEntities]
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.PhoneNumberEntities]
         """
         return self._phone_number_administration_client.phone_number_administration.get_all_releases(
             **kwargs
         )
 
-    @distributed_trace
-    def get_reservation_by_id(
-        self,
-        reservation_id,  # type: str
-        **kwargs  # type: Any
+    @distributed_trace_async
+    async def get_reservation_by_id(
+            self,
+            reservation_id,  # type: str
+            **kwargs  # type: Any
     ):
         # type: (...) -> PhoneNumberReservation
         """Get reservation by reservation id.
@@ -431,17 +440,18 @@ class PhoneNumberAdministrationClient(object):
         :type reservation_id: str
         :rtype: ~azure.communication.administration.PhoneNumberReservation
         """
-        return self._phone_number_administration_client.phone_number_administration.get_search_by_id(
+        return await self._phone_number_administration_client.phone_number_administration.get_search_by_id(
             search_id=reservation_id,
             **kwargs
         )
 
-    @distributed_trace
-    def begin_reserve_phone_numbers(
-        self,
-        **kwargs  # type: Any
+
+    @distributed_trace_async
+    async def begin_reserve_phone_numbers(
+            self,
+            **kwargs  # type: Any
     ):
-        # type: (...) -> LROPoller[PhoneNumberReservation]
+        # type: (...) -> AsyncLROPoller[PhoneNumberReservation]
         """Begins creating a phone number search to reserve phone numbers.
         Caller must provide one of the following:
          (1) all of keywords display_name, description, phone_plan_ids, area_code, quantity if all the phone plans
@@ -459,11 +469,11 @@ class PhoneNumberAdministrationClient(object):
         :keyword list[~azure.communication.administration.models.LocationOptionsDetails] location_options:
             the location options of the search.
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :rtype: ~azure.core.polling.LROPoller[~azure.communication.administration.PhoneNumberReservation]
+        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.communication.administration.PhoneNumberReservation]
         """
         cont_token = kwargs.pop('continuation_token', None)  # type: Optional[str]
 
-        reservation_polling = ReservePhoneNumberPolling(
+        reservation_polling = ReservePhoneNumberPollingAsync(
             is_terminated=lambda status: status in [
                 SearchStatus.Reserved,
                 SearchStatus.Expired,
@@ -474,7 +484,7 @@ class PhoneNumberAdministrationClient(object):
         )
 
         if cont_token is not None:
-            return LROPoller.from_continuation_token(
+            return AsyncLROPoller.from_continuation_token(
                 polling_method=reservation_polling,
                 continuation_token=cont_token,
                 client=self._phone_number_administration_client.phone_number_administration
@@ -498,16 +508,15 @@ class PhoneNumberAdministrationClient(object):
         if 'location_options' in kwargs:
             reservation_options.location_options = kwargs.pop('location_options')
 
-        create_reservation_response = self._phone_number_administration_client.\
+        create_reservation_response = await self._phone_number_administration_client.\
             phone_number_administration.create_search(
                 body=reservation_options,
                 **kwargs
         )
-
-        initial_state = self._phone_number_administration_client.phone_number_administration.get_search_by_id(
+        initial_state = await self._phone_number_administration_client.phone_number_administration.get_search_by_id(
             search_id=create_reservation_response.search_id
         )
-        return LROPoller(client=self._phone_number_administration_client.phone_number_administration,
+        return AsyncLROPoller(client=self._phone_number_administration_client.phone_number_administration,
                          initial_response=initial_state,
                          deserialization_callback=None,
                          polling_method=reservation_polling)
@@ -517,24 +526,24 @@ class PhoneNumberAdministrationClient(object):
             self,
             **kwargs  # type: Any
     ):
-        # type: (...) -> ItemPaged[PhoneNumberEntities]
+        # type: (...) -> AsyncItemPaged[PhoneNumberEntities]
         """Gets a list of all reservations.
 
         :keyword int skip: An optional parameter for how many entries to skip, for pagination purposes.
         The default is 0.
         :keyword int take: An optional parameter for how many entries to return, for pagination purposes.
         The default is 100.
-        :rtype: ~azure.core.paging.ItemPaged[~azure.communication.administration.PhoneNumberEntities]
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.communication.administration.PhoneNumberEntities]
         """
         return self._phone_number_administration_client.phone_number_administration.get_all_searches(
             **kwargs
         )
 
-    @distributed_trace
-    def cancel_reservation(
-        self,
-        reservation_id,  # type: str
-        **kwargs  # type: Any
+    @distributed_trace_async
+    async def cancel_reservation(
+            self,
+            reservation_id,  # type: str
+            **kwargs  # type: Any
     ):
         # type: (...) -> None
         """Cancels the reservation. This means existing numbers in the reservation will be made available.
@@ -543,28 +552,29 @@ class PhoneNumberAdministrationClient(object):
         :type reservation_id: str
         :rtype: None
         """
-        return self._phone_number_administration_client.phone_number_administration.cancel_search(
+        return await self._phone_number_administration_client.phone_number_administration.cancel_search(
             search_id=reservation_id,
             **kwargs
         )
 
-    @distributed_trace
-    def begin_purchase_reservation(
-        self,
-        **kwargs  # type: Any
+    @distributed_trace_async
+    async def begin_purchase_reservation(
+            self,
+            **kwargs  # type: Any
     ):
-        # type: (...) -> LROPoller[PhoneNumberReservation]
+
+        # type: (...) -> AsyncLROPoller[PhoneNumberReservation]
         """Begins purchase the reserved phone numbers of a phone number search.
         Caller must provide either reservation_id, or continuation_token keywords to use the method.
         If both reservation_id and continuation_token are specified, only continuation_token will be used to
         restart a poller from a saved state, and keyword reservation_id will be ignored.
         :keyword str reservation_id: The reservation id to be purchased.
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
-        :rtype: ~azure.core.polling.LROPoller[~azure.communication.administration.PhoneNumberReservation]
+        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.communication.administration.PhoneNumberReservation]
         """
         cont_token = kwargs.pop('continuation_token', None)  # type: Optional[str]
 
-        reservation_polling = PurchaseReservationPolling(
+        reservation_polling = PurchaseReservationPollingAsync(
             is_terminated=lambda status: status in [
                 SearchStatus.Success,
                 SearchStatus.Expired,
@@ -574,25 +584,37 @@ class PhoneNumberAdministrationClient(object):
         )
 
         if cont_token is not None:
-            return LROPoller.from_continuation_token(
+            return AsyncLROPoller.from_continuation_token(
                 polling_method=reservation_polling,
                 continuation_token=cont_token,
                 client=self._phone_number_administration_client.phone_number_administration
             )
 
-        if "reservation_id" not in kwargs:
+        reservation_id = kwargs.pop('reservation_id', None)  # type: str
+        if reservation_id is None:
             raise ValueError("Either kwarg 'reservation_id' or 'continuation_token' needs to be specified")
 
-        reservation_id = kwargs.pop('reservation_id')  # type: str
-
-        self._phone_number_administration_client.phone_number_administration.purchase_search(
-            search_id=reservation_id,
+        await self._phone_number_administration_client.phone_number_administration.purchase_search(
+            reservation_id,
             **kwargs
         )
-        initial_state = self._phone_number_administration_client.phone_number_administration.get_search_by_id(
+        initial_state = await self._phone_number_administration_client.phone_number_administration.get_search_by_id(
             search_id=reservation_id
         )
-        return LROPoller(client=self._phone_number_administration_client.phone_number_administration,
+        return AsyncLROPoller(client=self._phone_number_administration_client.phone_number_administration,
                          initial_response=initial_state,
                          deserialization_callback=None,
                          polling_method=reservation_polling)
+
+    async def __aenter__(self) -> "PhoneNumbersAdministrationClient":
+        await self._phone_number_administration_client.__aenter__()
+        return self
+
+    async def __aexit__(self, *args: "Any") -> None:
+        await self.close()
+
+    async def close(self) -> None:
+        """Close the :class:
+        `~azure.communication.administration.aio.PhoneNumbersAdministrationClient` session.
+        """
+        await self._phone_number_administration_client.__aexit__()